Move protocol definition to the right place
(services shouldn't be aware of Chalice)
