+++ src/azul/chalice.py
@@ -17,7 +17,6 @@ from typing import (
     Callable,
     Iterator,
     Literal,
-    Protocol,
     Self,
     Sequence,
 )
@@ -894,12 +893,3 @@ class AppController:
     def current_request(self) -> AzulRequest:
         assert self.app.current_request is not None
         return self.app.current_request
-
-
-class SchemaUrlFunc(Protocol):
-
-    def __call__(self,
-                 *,
-                 schema_name: str,
-                 version: int
-                 ) -> mutable_furl: ...
